update motor cmd

+++ ui/src/lightning_gui/src/lightning_gui_main/pc_autox/pc_data.h
@@ -195,9 +195,9 @@ typedef struct
 typedef struct 
 {
 	MotorMsgHeader header;
-	uint16_t     speed;
-	uint16_t     angle;
-	uint16_t     location;
+	float     speed;
+	float     angle;
+	float     location;
 	uint8_t		mode;
 	MotorMsgTailer 	tailer; 
 } motorWorkModeMsg;
